refactor: separated function for clearing
